RequirementMachine: Split off Symbol.cpp and Term.cpp from RewriteSystem.cpp
