+++ lib/AST/RequirementMachine/RewriteSystemCompletion.cpp
@@ -26,7 +26,9 @@
 //===----------------------------------------------------------------------===//
 
 #include "swift/Basic/Defer.h"
+#include "swift/Basic/Range.h"
 #include "llvm/ADT/FoldingSet.h"
+#include "llvm/Support/Debug.h"
 #include "llvm/Support/raw_ostream.h"
 #include <algorithm>
 #include <deque>
